BREAKING CHANGE: enforce code coverage and test suite failures  (#729)
* BREAKING CHANGE: enforce code coverage

Previously code coverage seems to have not been enforced.
To add to the hilarity, no one seems to have caught on, or
cared. Maybe they are wiser than me, because code coverage seems
to be a very blunt tool which forces people to care about its
letter rather than the intent, and usually makes people write
silly pointless tests, probably because the coverage numbers are
set too high.

Anyway, more to the point, runCLI is an undocumented API. Hence
this is not even the right fix. The right fix would be to stop
using an undocumented API, and instead use it as it was designed,
which I think means via shell script. I'm not pursuing that out
of time constraints, but I see it as far superior. Using
undocumented features makes it more likely that they will be used
incorrectly (as was being done here.). Also -- why wasn't this
ever tested to see that it worked? Maybe I'm missing something
super obvious here :|

I did not rely on any documentation for this fix, because there
are no official docs, but it seems like this is the right solution,
and it was very easy to test.

This is a breaking change because any repo which is using ds-cli
and ignoring code coverage -- probably most of them -- will stop
working once this change is made. If ds-cli were not in use by anyone,
I would call it a patch update, but as it is, this instantly breaks
any consumers that think their coverage is working and never looked
at their logs. So I think a major version update is warranted, as
a warning.

* feat:  ignore .idea; specify that node-gyp errors are not blocking

-Ignore JetBrains IDE-specific files
-Specify that even if Node-gyp throws out errors (... errors that
do not end in "ok" :/ ), the build may still work.